VectorGraphicsEditor/requirements.txt
asgiref==3.7.2
canvasvg==1.0.5
Pillow==9.5.0
six==1.16.0
tzdata==2023.3
